Crate example_cli

Re-exports

pub use anyhow;

pub use bdk_file_store;

pub use clap;

Structs

Args

CreateTxChange

Init

The initial state returned by init.

Enums

AddressCmd

CoinSelectionAlgo

Commands

Keychain

TxOutCmd

Functions

create_tx

handle_commands

init

Parses command line arguments and initializes all components, creating a file store with the given parameters, or loading one if it exists.

planned_utxos

Type Definitions

Database

KeychainChangeSet

KeychainTxGraph

PlannedUtxo